Fix _Spy: true segfault on ARM64 macOS (64-bit pointer truncation)
set_color(int) truncated 64-bit CGColorRef pointers to 32 bits,
causing CGColorEqualToColor to dereference invalid memory. Changed
parameter type to long int to match the color accessor return types.

Broke platformWindow.hh symlink to vm/ since the 32-bit VM uses int.

+++ vm64/src/any/os/quartzWindow.mm
@@ -948,9 +948,9 @@ void QuartzWindow::fill_rectangle(int x, int y, int w, int h) {
   CGContextFillRect(myContext, CGRectMake(x, y-1, w, h+1));
 }
 
-void QuartzWindow::set_color(int c) {
-  CGContextSetFillColorWithColor(   myContext, (CGColorRef)(long)c );
-  CGContextSetStrokeColorWithColor( myContext, (CGColorRef)(long)c );
+void QuartzWindow::set_color(long int c) {
+  CGContextSetFillColorWithColor(   myContext, (CGColorRef)c );
+  CGContextSetStrokeColorWithColor( myContext, (CGColorRef)c );
 }
 
 void QuartzWindow::set_thickness(int t) {
